Переименует типы доп опций
