+++ src/view/offer-available-template.js
