+++ src/view/offer-selected-template.js
