+++ src/view/offer-selected-view.js
@@ -1,8 +1,8 @@
 import BaseView from './base-view.js';
-import createPointOfferTemplate from './point-offer-template.js';
+import createOfferSelectedTemplate from './offer-selected-template.js';
 
 /** Представление (выбранной) дополнительной опции */
-export default class PointOfferView extends BaseView {
+export default class OfferSelectedView extends BaseView {
   constructor() {
     super();
 
@@ -11,7 +11,7 @@ export default class PointOfferView extends BaseView {
 
   /** @override */
   createAdjacentHtml() {
-    return createPointOfferTemplate();
+    return createOfferSelectedTemplate();
   }
 
   /**
@@ -39,4 +39,4 @@ export default class PointOfferView extends BaseView {
   }
 }
 
-customElements.define('offer-selected', PointOfferView);
+customElements.define('offer-selected', OfferSelectedView);
